backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2

# Development
python-dotenv==1.0.0
